implemented getting random recipe

+++ FoodAssistant/Server/src/main/kotlin/ru/spbau/shavkunov/server/data/Ingredient.kt
@@ -1,4 +1,4 @@
 package ru.spbau.shavkunov.server.data
 
-class Ingredient(var name: String?, var amount: Int?, var kindOfAmount: AmountType?) {
+class Ingredient(var name: String?, var amount: Double?, var kindOfAmount: AmountType?) {
 }
